[TDC-143] UserCoupon Domain 보완
