+++ threedollar-core/threedollar-domain/src/main/java/com/threedollar/domain/coupon/CouponGroup.java
@@ -5,6 +5,9 @@ import lombok.Getter;
 @Getter
 public enum CouponGroup {
     BOSS_STORE("사장님 가게")
+    ,ADVERTISEMENT_COUPON("광고 쿠폰")
+    ,EVENT("이벤트 쿠폰")
+    ,DISCOUNT("할인 쿠폰")
     ;
     private final String description;
     CouponGroup(String description) {
